+++ doc_newmaingateway_http_test.go
@@ -10,11 +10,11 @@ import (
 
 // NOTE: your application MUST use `package main` and define a `main()` function.  The
 // example text is to make the documentation compatible with godoc.
-
 func echoAPIGatewayHTTPEvent(ctx context.Context,
 	props map[string]interface{}) error {
 	lambdaCtx, _ := lambdacontext.FromContext(ctx)
-	Logger().WithFields(logrus.Fields{
+	logger, _ := ctx.Value(ContextKeyLogger).(*logrus.Logger)
+	logger.WithFields(logrus.Fields{
 		"RequestID":  lambdaCtx.AwsRequestID,
 		"Properties": props,
 	}).Info("Lambda event")